Vector homework
Write a constructor called Vector that will represent a point on an x,y-axis, like a graph. Have it accept 2 parameters, x and y.

Have the Vector prototype include two methods plus and minus that accepts a Vector object as their parameter and have them return a new Vector object that is the sum or difference of those points.

Lastly, add a length getter that returns the distance from the coordinates (0,0) to (x,y).